refactor: use float consistently for coordinates

+++ include/robots/Robot.h
@@ -60,18 +60,18 @@ class Robot {
     void addSensor(std::unique_ptr<Sensor> sensor);
     void updateSensors();
 
-    [[nodiscard]] double x() const noexcept;
-    [[nodiscard]] double y() const noexcept;
-    [[nodiscard]] double angle() const noexcept;
+    [[nodiscard]] float x() const noexcept;
+    [[nodiscard]] float y() const noexcept;
+    [[nodiscard]] float angle() const noexcept;
 
   protected:
-    void moveForward(double distance);
-    void rotate(double degree);
+    void moveForward(float distance);
+    void rotate(float degree);
 
   private:
-    double x_;
-    double y_;
-    double angle_;
+    float x_;
+    float y_;
+    float angle_;
     float speed_;
     Vector2 targetPosition_;
     float rotationSpeed_;
